Don't log ECONNRESET and EPIPE when responding to the client
If we send a response that's too big for the client's buffer, the
client will disconnect and not read the rest of our response, and
then come back with a new connection after increasing its buffer.
There's no need to log that, and generally, clients can disappear at
any point.