Use fixed pool of threads to handle incoming requests
Uses a fixed pool of threads to handle incoming requests. Threads pass a
UnixStream over a channel to each other. If no worker is available
inside the libc timeout for nscd, forces an exit.

Adds a WorkGroup to coordinate the pool of threads which allows blocking
until any of the worker threads panic or exit normally.

Closes #24