Implement the GETAI operation
We implement the GETAI (aka. get address info) operation. Most of the
glibc applications will rely on this operation to perform the hostname
-> addrs resolutions.

We add some unit tests checking that we're correctly serializing the
address info header and its payload. The expected data used in these
tests has been extracted from various Nscd socket dumps. We also add
an integration test checking that getaddrinfo is able to resolve
localhost.

The Nix crate we rely on for the glibc FFI bindings does not wrap
getaddrinfo yet. We had to pull the `dns-lookup` package to get those
bindings.