Getaddrinfo: always retrieve the canonical name
Nscd always sets the AI_CANONNAME flag for a getaddrinfo request. When
this flag is on, the canonical name (~ FQDN) for the requested address is
retrieved.

We found this issue through the nixosTests.hostname.explicitDomain
NixOS VM test. It went unnoticed in the wild probably because
the nscd client tend to fill canonical name in the request itself
once it retrieved it once.

While investigating this issue, I realized that setting the
SOCK_STREAM flag gets rid of the duplicate addrs. Meaning that we do
not need to filter them ourselves with a HashSet anymore.