serialize_hostent: serialize IP addresses in one go
Copying the IP adresses to the target addresses buffer in one go
instead of copying them byte by byte.

We're already doing that for the getaddrinfo requests. It seems like
we forgot to backport this to the legacy hostent operations.

I stumbled upon that while profiling a system doing a lot of
gethostbyname requests. Nsncd was taking ~60% of the overall system
load on it, mostly iterating on the IP adresses octets. The load is
back to a more normal level once this diff is applied.