Remove vendor rules target
Builds started to fail recently on all debian versions due to an issue
with the vendor target used for offline builds.

The closest reported issue I could find was on openwrt; related to a
rustlang change, both are referenced below.

https://github.com/openwrt/openwrt/issues/20189

https://github.com/rust-lang/cargo/commit/29ecb7f36a963bed4729bd6909200b4302a366a5

We can skip the vendor rules as this build can access crates.io, this
target was designed for offline builds at the original authors
organization.

Once debian 11 is retired, debian/control will be updated to use
dependencies packages by the distribution as debian upstream does